requirements.txt
pillow
torch==2.3.0
torchvision==0.18.0
fastapi==0.111.0
uvicorn==0.19.0
pickle5==0.0.11
scikit-learn